refactored fitting stuff out of DatAttribute
