+++ src/DatObject/Attributes/DatAttribute.py
@@ -1,22 +1,19 @@
 from __future__ import annotations
-import pandas as pd
 import datetime
 import os
-from hashlib import md5
-import inspect
 from typing import Union, List, Optional, TypeVar, Type, Callable, Any, Dict, Tuple
-from src.HDF_Util import is_DataDescriptor, NotFoundInHdfError, is_Group
+
+from src.AnalysisTools.fitting import Values, FitInfo, FitIdentifier, calculate_fit
+from src.HDF_Util import NotFoundInHdfError, is_Group
 from src import CoreUtil as CU
 import abc
 import h5py
 import logging
 import numpy as np
 import lmfit as lm
-from dataclasses import dataclass, field, InitVar
-from src.HDF_Util import params_from_HDF, params_to_HDF, with_hdf_read, with_hdf_write
+from dataclasses import dataclass, field
+from src.HDF_Util import with_hdf_read, with_hdf_write
 import src.HDF_Util as HDU
-from functools import lru_cache, partial
-from src.CoreUtil import MyLRU, my_partial
 from deprecation import deprecated
 from typing import TYPE_CHECKING
 
@@ -37,7 +34,7 @@ def update_meta(self, other):
     return self
 
 
-class LateBindingProperty (property):
+class LateBindingProperty(property):
     """https://code.activestate.com/recipes/408713-late-binding-properties-allowing-subclasses-to-ove/
 
     """
@@ -400,7 +397,8 @@ class DatDataclassTemplate(abc.ABC):
         # Add additional things from HDF
         d = dict(**d, **cls.additional_load_from_hdf(dc_group))
 
-        d = {k: v if not isinstance(v, h5py.Dataset) else v[:] for k, v in d.items()}  # Load all data into memory here if necessary
+        d = {k: v if not isinstance(v, h5py.Dataset) else v[:] for k, v in
+             d.items()}  # Load all data into memory here if necessary
         inst = cls(**d)
         return inst
 
@@ -408,7 +406,8 @@ class DatDataclassTemplate(abc.ABC):
         ignore_keys = CU.ensure_set(ignore_keys)
         for k in set(self.__annotations__) - ignore_keys:
             val = getattr(self, k)
-            if isinstance(val, (np.ndarray, h5py.Dataset)) and val.size > 1000:  # Pretty much anything that is an array should be saved as a dataset
+            if isinstance(val, (np.ndarray,
+                                h5py.Dataset)) and val.size > 1000:  # Pretty much anything that is an array should be saved as a dataset
                 HDU.set_data(group, k, val)
             elif HDU.allowed(val):
                 HDU.set_attr(group, k, val)
@@ -435,58 +434,6 @@ class DatDataclassTemplate(abc.ABC):
         return cls.__name__
 
 
-class Values(object):
-    """Object to store Init/Best values in and stores Keys of those values in self.keys"""
-
-    def __init__(self):
-        self.keys = []
-
-    def __getattr__(self, item):
-        if item.startswith('__') or item.startswith(
-                '_') or item == 'keys':  # So don't complain about things like __len__
-            return super().__getattribute__(item)  # Come's here looking for Ipython variables
-        else:
-            if item in self.keys:
-                return super().__getattribute__(item)
-            else:
-                msg = f'{item} does not exist. Valid keys are {self.keys}'
-                print(msg)
-                logger.warning(msg)
-                return None
-
-    def get(self, item, default=None):
-        if item in self.keys:
-            val = self.__getattr__(item)
-        else:
-            val = default
-        if default is not None and val is None:
-            return default
-        return val
-
-    def __setattr__(self, key, value):
-        if key.startswith('__') or key.startswith('_') or key == 'keys' or not isinstance(value, (
-                np.number, float, int, type(None))):  # So don't complain about
-            # things like __len__ and don't keep key of random things attached to class
-            super().__setattr__(key, value)
-        else:  # probably is something I want the key of
-            self.keys.append(key)
-            super().__setattr__(key, value)
-
-    def __repr__(self):
-        string = ''
-        for key in self.keys:
-            v = getattr(self, key)
-            if v is not None:
-                string += f'{key}={self.__getattr__(key):.5g}\n'
-            else:
-                string += f'{key}=None\n'
-        return string
-
-    def to_df(self):
-        df = pd.DataFrame(data=[[self.get(k) for k in self.keys]], columns=[k for k in self.keys])
-        return df
-
-
 @dataclass
 class NewFitInfo(DatDataclassTemplate):
     params: lm.Parameters
@@ -497,179 +444,6 @@ class NewFitInfo(DatDataclassTemplate):
     model: lm.Model
     best_values: Values
     init_values: Values
-
-
-@dataclass
-class FitInfo(DatDataclassTemplate):
-    params: Union[lm.Parameters, None] = None
-    init_params: lm.Parameters = None
-    func_name: Union[str, None] = None
-    func_code: Union[str, None] = None
-    fit_report: Union[str, None] = None
-    model: Union[lm.Model, None] = None
-    best_values: Union[Values, None] = None
-    init_values: Union[Values, None] = None
-    success: bool = None
-    hash: int = None
-
-    # Will only exist when set from fit, or after recalculate_fit
-    fit_result: Union[lm.model.ModelResult, None] = None
-
-    def init_from_fit(self, fit: lm.model.ModelResult, hash_: Optional[int] = None):
-        """Init values from fit result"""
-        if fit is None:
-            logger.warning(f'Got None for fit to initialize from. Not doing anything.')
-            return None
-        assert isinstance(fit, lm.model.ModelResult)
-        self.params = fit.params
-        self.init_params = fit.init_params
-        self.func_name = fit.model.func.__name__
-
-        #  Can't get source code when running from deepcopy (and maybe other things will break this)
-        try:
-            func_code = inspect.getsource(fit.model.func)
-        except OSError:
-            if self.func_code is not None:
-                func_code = '[WARNING]: might not be correct as fit was re run and could not get source code: ' + self.func_code
-            else:
-                logger.warning('Failed to get source func_code and no existing func_code')
-                func_code = 'Failed to get source code due to OSError'
-        self.func_code = func_code
-
-        self.fit_report = fit.fit_report()
-        self.success = fit.success
-        self.model = fit.model
-        self.best_values = Values()
-        self.init_values = Values()
-        for key in self.params.keys():
-            par = self.params[key]
-            self.best_values.__setattr__(par.name, par.value)
-            self.init_values.__setattr__(par.name, par.init_value)
-        self.hash = hash_
-        self.fit_result = fit
-
-    def init_from_hdf(self, group: h5py.Group):
-        """Init values from HDF file"""
-        self.params = params_from_HDF(group)
-        self.init_params = params_from_HDF(group.get('init_params'), initial=True)
-        self.func_name = group.attrs.get('func_name', None)
-        self.func_code = group.attrs.get('func_code', None)
-        self.fit_report = group.attrs.get('fit_report', None)
-        self.model = lm.models.Model(self._get_func())
-        self.success = group.attrs.get('success', None)
-
-        self.best_values = Values()
-        self.init_values = Values()
-        for key in self.params.keys():
-            par = self.params[key]
-            self.best_values.__setattr__(par.name, par.value)
-            self.init_values.__setattr__(par.name, par.init_value)
-
-        temp_hash = group.attrs.get('hash')
-        if temp_hash is not None:
-            self.hash = int(temp_hash)
-        else:
-            self.hash = None
-        self.fit_result = None
-
-    def save_to_hdf(self, parent_group: h5py.Group, name: Optional[str] = None):
-        if name is None:
-            name = self._default_name()
-        parent_group = parent_group.require_group(name)
-
-        if self.params is None:
-            logger.warning(f'No params to save for {self.func_name} fit. Not doing anything')
-            return None
-        params_to_HDF(self.params, parent_group)
-        params_to_HDF(self.init_params, parent_group.require_group('init_params'))
-        parent_group.attrs['description'] = 'FitInfo'  # Overwrites what params_to_HDF sets
-        parent_group.attrs['func_name'] = self.func_name
-        parent_group.attrs['func_code'] = self.func_code
-        parent_group.attrs['fit_report'] = self.fit_report
-        parent_group.attrs['success'] = self.success
-        if self.hash is not None:
-            parent_group.attrs['hash'] = int(self.hash)
-
-    def _get_func(self):
-        """Did not initially enforce having a good way to get back the lm.model when loading from hdf, so this is
-        the workaround... Let FitInfo know what the functions are for the saved function name.
-        Note: name part must match function name exactly"""
-        # return HDU.get_func(self.func_name, self.func_code)
-        from src.DatObject.Attributes.Transition import i_sense, i_sense_strong, i_sense_digamma, i_sense_digamma_quad, i_sense_digamma_amplin
-        from src.DatObject.Attributes.Entropy import entropy_nik_shape
-        funcs = {
-            'i_sense': i_sense,
-            'i_sense_strong': i_sense_strong,
-            'i_sense_digamma': i_sense_digamma,
-            'i_sense_digamma_quad': i_sense_digamma_quad,
-            'i_sense_digamma_amplin': i_sense_digamma_amplin,
-            'entropy_nik_shape': entropy_nik_shape,
-        }
-        if self.func_name in funcs:
-            return funcs[self.func_name]
-        else:
-            raise KeyError(f'{self.func_name} not a recongnized function in '
-                           f'src.DatObject.Attributes.DatAttribute.FitInfo')
-
-    def eval_fit(self, x: np.ndarray):
-        """Return best fit for x array using params"""
-        return self.model.eval(self.params, x=x)
-
-    def eval_init(self, x: np.ndarray):
-        """Return init fit for x array using params"""
-        init_pars = CU.edit_params(self.params, list(self.params.keys()),
-                                   [par.init_value for par in self.params.values()])
-        return self.model.eval(init_pars, x=x)
-
-    def recalculate_fit(self, x: np.ndarray, data: np.ndarray, auto_bin=False):
-        """Fit to data with x array and update self"""
-        assert data.ndim == 1
-        data, x = CU.remove_nans(data, x)
-        if auto_bin is True and len(data) > FIT_NUM_BINS:
-            logger.info(f'Binning data of len {len(data)} into {FIT_NUM_BINS} before fitting')
-            x, data = CU.bin_data([x, data], round(len(data) / FIT_NUM_BINS))
-        fit = self.model.fit(data.astype(np.float32), self.params, x=x, nan_policy='omit')
-        self.init_from_fit(fit, self.hash)
-
-    def edit_params(self, param_names=None, values=None, varys=None, mins=None, maxs=None):
-        self.params = CU.edit_params(self.params, param_names, values, varys, mins, maxs)
-
-    def to_df(self):
-        val_df = self.best_values.to_df()
-        val_df['success'] = self.success
-        return val_df
-
-    def __hash__(self):
-        if self.hash is None:
-            raise AttributeError(f'hash value stored as None so hashing not supported')
-        return int(self.hash)
-
-    def __eq__(self, other):
-        if isinstance(other, self.__class__):
-            return hash(other) == hash(self)
-        return False
-
-    def __repr__(self):
-        return self.fit_report
-
-    @classmethod
-    def from_fit(cls, fit, hash_: Optional[int] = None):
-        """Use FitIdentifier to generate hash (Should be done before binning data to be able to check if
-        matches before doing expensive processing)"""
-        inst = cls()
-        inst.init_from_fit(fit, hash_)
-        return inst
-
-    @classmethod
-    def from_hdf(cls, parent_group: h5py.Group, name: str = None):
-        if name is None:
-            name = cls._default_name()
-        fg = parent_group.get(name)
-        if fg is None:
-            raise NotFoundInHdfError(f'{name} not found in {parent_group.name}')
-        inst = cls()
-        inst.init_from_hdf(fg)
-        return inst
 
 
 class DatAttributeWithData(DatAttribute, abc.ABC):
@@ -734,56 +508,6 @@ class DatAttributeWithData(DatAttribute, abc.ABC):
         return D.get_data_descriptor(name, filled=filled, data_group_name=self.group_name)
 
 
-@dataclass
-class FitIdentifier:
-    initial_params: lm.Parameters
-    func: Callable  # Or should I just use func name here? Or func code?
-    data: InitVar[np.ndarray]
-    data_hash: str = field(init=False)
-
-    def __post_init__(self, data: np.ndarray):
-        assert isinstance(self.initial_params, lm.Parameters)
-        self.data_hash = self._hash_data(data)
-
-    @staticmethod
-    def _hash_data(data: np.ndarray):
-        if data.ndim == 1:
-            data = data[~np.isnan(data)]  # Because fits omit NaN data so this will make data match the fit data.
-        return md5(data.tobytes()).hexdigest()
-
-    def __hash__(self):
-        """The default hash of FitIdentifier which will allow comparison between instances
-        Using hashlib hashes makes this deterministic rather than runtime specific, so can compare to saved values
-        """
-        pars_hash = self._hash_params()
-        func_hash = self._hash_func()
-        data_hash = self.data_hash
-        h = md5(pars_hash.encode())
-        h.update(func_hash.encode())
-        h.update(data_hash.encode())
-        return int(h.hexdigest(), 16)
-
-    def __eq__(self, other):
-        if isinstance(other, self.__class__):
-            if hash(self) == hash(other):
-                return True
-        return False
-
-    def _hash_params(self) -> str:
-        h = md5(str(sorted(self.initial_params.valuesdict().items())).encode())
-        return h.hexdigest()
-
-    def _hash_func(self) -> str:
-        # hash(self.func)   # Works pretty well, but if function is executed later even if it does the same thing it
-        # will change
-        h = md5(str(self.func.__name__).encode())
-        return h.hexdigest()
-
-    def generate_name(self):
-        """ Will be some thing reproducible and easy to read. Note: not totally guaranteed to be unique."""
-        return str(hash(self))[0:5]
-
-
 # def get_all_fit_paths(group: h5py.Group) -> List[str]:
 #     fit_paths = []
 #     for k in group.keys():
@@ -809,8 +533,8 @@ class FitPaths:
         def get_paths_in_group(group: h5py.Group) -> List[str]:
             if (paths := group.attrs.get('all_paths', None)) is None:
                 paths = HDU.find_all_groups_names_with_attr(group,
-                                                        attr_name='description',
-                                                        attr_value='FitInfo')
+                                                            attr_name='description',
+                                                            attr_value='FitInfo')
             return paths
 
         avg_fit_paths = get_paths_in_group(avg_fit_group)
@@ -964,7 +688,7 @@ class FittingAttribute(DatAttributeWithData, DatAttribute, abc.ABC):
     def avg_data(self):
         """Quick access for DEFAULT avg_data ONLY"""
         if self._avg_data is None:
-             self._avg_data, self._avg_data_std, self._avg_x = self.get_avg_data(x=self.x, data=self.data,
+            self._avg_data, self._avg_data_std, self._avg_x = self.get_avg_data(x=self.x, data=self.data,
                                                                                 centers=None, return_x=True,
                                                                                 return_std=True)
         return self._avg_data
@@ -973,7 +697,7 @@ class FittingAttribute(DatAttributeWithData, DatAttribute, abc.ABC):
     def avg_x(self):
         """Quick access for DEFAULT avg_x ONLY (although this likely be the same all the time)"""
         if self._avg_x is None:
-             self._avg_data, self._avg_data_std, self._avg_x = self.get_avg_data(x=self.x, data=self.data,
+            self._avg_data, self._avg_data_std, self._avg_x = self.get_avg_data(x=self.x, data=self.data,
                                                                                 centers=None, return_x=True,
                                                                                 return_std=True)
         return self._avg_x
@@ -982,9 +706,9 @@ class FittingAttribute(DatAttributeWithData, DatAttribute, abc.ABC):
     def avg_data_std(self):
         """Quick access for DEFAULT avg_data_std ONLY"""
         if self._avg_data_std is None:
-             self._avg_data, self._avg_data_std, self._avg_x = self.get_avg_data(x=self.x, data=self.data,
-                                                                            centers=None, return_x=True,
-                                                                            return_std=True)
+            self._avg_data, self._avg_data_std, self._avg_x = self.get_avg_data(x=self.x, data=self.data,
+                                                                                centers=None, return_x=True,
+                                                                                return_std=True)
         return self._avg_data_std
 
     @property
@@ -1025,16 +749,17 @@ class FittingAttribute(DatAttributeWithData, DatAttribute, abc.ABC):
         """
         # Try to get saved avg_data and avg_x
         if not name:
-            avg_data_name = self.DEFAULT_DATA_NAME+'_avg'
+            avg_data_name = self.DEFAULT_DATA_NAME + '_avg'
             avg_x_name = 'x_avg'
         else:
-            avg_data_name = name+'_avg'
+            avg_data_name = name + '_avg'
             avg_x_name = f'x_avg_for_{name}'
 
-        if all([v in self.specific_data_descriptors_keys.keys() for v in [avg_data_name, avg_x_name, avg_data_name + '_std']]):
+        if all([v in self.specific_data_descriptors_keys.keys() for v in
+                [avg_data_name, avg_x_name, avg_data_name + '_std']]):
             avg_data = self.get_data(avg_data_name)
             avg_x = self.get_data(avg_x_name)
-            avg_data_std = self.get_data(avg_data_name+'_std')
+            avg_data_std = self.get_data(avg_data_name + '_std')
         else:
             # Otherwise create avg_data and avg_x
             if x is None:
@@ -1046,7 +771,7 @@ class FittingAttribute(DatAttributeWithData, DatAttribute, abc.ABC):
             avg_x, avg_data, avg_data_std = self._make_avg_data(x, data, centers)
             self.set_data(avg_x_name, avg_x)
             self.set_data(avg_data_name, avg_data)
-            self.set_data(avg_data_name+'_std', avg_data_std)
+            self.set_data(avg_data_name + '_std', avg_data_std)
 
         ret = [avg_data]
         if return_std:
@@ -1057,7 +782,8 @@ class FittingAttribute(DatAttributeWithData, DatAttribute, abc.ABC):
             ret = ret[0]
         return ret
 
-    def _make_avg_data(self, x: np.ndarray, data: np.ndarray, centers: Optional[List[float]] = None) -> Tuple[np.ndarray, np.ndarray, np.ndarray]:
+    def _make_avg_data(self, x: np.ndarray, data: np.ndarray, centers: Optional[List[float]] = None) -> Tuple[
+        np.ndarray, np.ndarray, np.ndarray]:
         """
         Generates average data using inputs and returns avg_x, avg_data, avg_data_std
         Args:
@@ -1126,7 +852,8 @@ class FittingAttribute(DatAttributeWithData, DatAttribute, abc.ABC):
                 fit_path = self._get_fit_path_from_name(name, which, row)
                 if fit_path:  # If found get fit
                     fit = self._get_fit_from_path(fit_path)
-                    if not any((initial_params, fit_func, data is not None)) or check_exists:  # If nothing to compare to or ONLY looking for existing
+                    if not any((initial_params, fit_func,
+                                data is not None)) or check_exists:  # If nothing to compare to or ONLY looking for existing
                         return fit
                 elif check_exists:
                     raise NotFoundInHdfError(f'{name} not found for dat{self.dat.datnum} in {self.group_name}')
@@ -1134,7 +861,8 @@ class FittingAttribute(DatAttributeWithData, DatAttribute, abc.ABC):
             # Should ONLY get past here with check_exists == True if name is None
             if check_exists:
                 if name is not None:
-                    raise RuntimeError(f'Dat{self.dat.datnum}: should not have got here with name={name} and check_exists={check_exists}')
+                    raise RuntimeError(
+                        f'Dat{self.dat.datnum}: should not have got here with name={name} and check_exists={check_exists}')
 
             # Special name default if nothing else specified
             if not name and not any((initial_params, fit_func, data is not None)):
@@ -1178,18 +906,22 @@ class FittingAttribute(DatAttributeWithData, DatAttribute, abc.ABC):
                 if hash(fit) == hash(fit_id):
                     if name and fit_path and name not in fit_path:
                         if check_exists:
-                            raise NotFoundInHdfError(f'{name} was not found, although a fit with the same parameters WAS found at {fit_path}')
-                        logger.warning(f'Asked for {name} but fit already exists at {fit_path}. A duplicate will be saved')
+                            raise NotFoundInHdfError(
+                                f'{name} was not found, although a fit with the same parameters WAS found at {fit_path}')
+                        logger.warning(
+                            f'Asked for {name} but fit already exists at {fit_path}. A duplicate will be saved')
                         self._save_fit(fit, which, name, row=row)
                     return fit
                 else:
-                    logger.warning(f'Fit found with same initial arguments, but hash does not match. Recalculating fit now')
+                    logger.warning(
+                        f'Fit found with same initial arguments, but hash does not match. Recalculating fit now')
 
             # Otherwise start generating new fit
             if not name:  # Generate anything other than default name
                 name = fit_id.generate_name()
 
-        fit = self._calculate_fit(x, data, params=initial_params, func=fit_func, auto_bin=True)
+        fit = self._calculate_fit(x, data, params=initial_params, func=fit_func, auto_bin=True,
+                                  generate_hash=True if not calculate_only else False)
         if fit and not calculate_only:
             self._save_fit(fit, which, name, row=row)
         return fit
@@ -1257,24 +989,24 @@ class FittingAttribute(DatAttributeWithData, DatAttribute, abc.ABC):
             (str): Name to save fit as (or to find fit as)
         """
         if which == 'avg':
-            return name+'_avg'
+            return name + '_avg'
         elif which == 'row':
-            return name+f'_row[{row}]'
+            return name + f'_row[{row}]'
         else:
             raise ValueError(f'{which} not in ["avg", "row"]')
 
     def _get_fit_parent_group_name(self, which: str, row: int = 0) -> str:
         """Get path to parent group of avg or row fit"""
         if which == 'avg':
-            group_name = '/'+'/'.join((self.group_name, 'Avg Fits'))
+            group_name = '/' + '/'.join((self.group_name, 'Avg Fits'))
         elif which == 'row':
-            group_name = '/'+'/'.join((self.group_name, 'Row Fits', str(row)))
+            group_name = '/' + '/'.join((self.group_name, 'Row Fits', str(row)))
         else:
             raise ValueError(f'{which} not in ["avg", "row"]')
         return group_name
 
     def _calculate_fit(self, x: np.ndarray, data: np.ndarray, params: lm.Parameters, func: Callable[[Any], float],
-                       auto_bin=True) -> FitInfo:
+                       auto_bin=True, generate_hash: bool = True) -> FitInfo:
         """
         Calculates fit on data (Note: assumes that 'x' is the independent variable in fit_func)
         Args:
@@ -1284,33 +1016,13 @@ class FittingAttribute(DatAttributeWithData, DatAttribute, abc.ABC):
             func (Callable): Function to fit to
             auto_bin (bool): if True will bin data into self.AUTO_BIN_SIZE if data has more data points (can massively
             increase computation speed without noticeable change to fit values for ~1000)
+            generate_hash: Generate hash before binning data so that a future call can be compared before calculation
 
         Returns:
             (FitInfo): FitInfo instance (with FitInfo.fit_result filled)
         """
-        model = lm.model.Model(func)
-        hash_ = hash(FitIdentifier(params, func, data))  # Needs to be done BEFORE binning data.
-        if auto_bin and data.shape[-1] > self.AUTO_BIN_SIZE:
-            bin_size = int(np.floor(data.shape[-1] / self.AUTO_BIN_SIZE))
-            x, data = [CU.bin_data_new(arr, bin_x=bin_size) for arr in [x, data]]
-        for par in params:
-            params[par].value = np.float32(params[par].value)  # VERY infrequently causes issues for calculating uncertainties with np.float64 dtype
-        try:
-            fit = FitInfo.from_fit(model.fit(data.astype(np.float32), params, x=x.astype(np.float32), nan_policy='omit'), hash_)
-            if fit.fit_result.covar is None and fit.success is True:  # Failed to calculate uncertainties even though fit was successful
-                logger.warning(f'Dat{self.dat.datnum}: Uncertainties failed')
-            elif fit.success is False:
-                logger.warning(f'Dat{self.dat.datnum}: A fit failed')
-                # logger.warning(f'Dat{self.dat.datnum}: Uncertainties failed, trying with 10% starting estimate')
-                # fit = fit.fit_result
-                # for p in fit.params:  # Set 10% as initial estimate
-                #     fit.params[p].stderr = abs(fit.params[p].value*0.1)
-                # fit.conf_interval()  # Try calculate conf intervals again
-                # fit = FitInfo.from_fit(fit, hash_)
-        except TypeError as e:
-            logger.warning(f'{e} while fitting in {self.group_name} for {self.dat.dat_id}')
-            fit = None
-        return fit
+        return calculate_fit(x=x, data=data, params=params, func=func, auto_bin=auto_bin, min_bins=self.AUTO_BIN_SIZE,
+                             generate_hash=generate_hash, warning_id=f'Dat{self.dat.datnum}')
 
     @with_hdf_write
     def initialize_minimum(self):
@@ -1494,4 +1206,5 @@ class DataDescriptor(DatDataclassTemplate):
             return np.s_[:]
         else:
             raise NotImplementedError(f'Still need to write how to get slice of only good rows! ')  # TODO: Do this
+
